Cambios a user_updated para vercel
